requirements.txt
transformers>=4.43.0
torch
huggingface_hub
faiss-cpu
sentence-transformers